(feat): init library setup TokenLibrary:  - added tokens  - added publish plugin  ComponentLibrary:  - added publish plugin  - added components  - api import of token lib
