+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
         android:allowBackup="true"
@@ -10,12 +9,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MobileDesignSystem"
-        tools:targetApi="31">
+        android:theme="@style/Theme.MobileDesignSystem">
         <activity
-            android:name=".MainActivity"
+            android:name="com.paixao.dev.mobile_design_system.ui.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.MobileDesignSystem">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
