--- app/src/main/java/com/paixao/dev/mobile_design_system/ui/theme/Color.kt
@@ -1,61 +1,0 @@
-package com.paixao.dev.mobile_design_system.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val primaryLight = Color(0xFF415F91)
-val onPrimaryLight = Color(0xFFFFFFFF)
-val primaryContainerLight = Color(0xFFD6E3FF)
-val onPrimaryContainerLight = Color(0xFF001B3E)
-val secondaryLight = Color(0xFF565F71)
-val onSecondaryLight = Color(0xFFFFFFFF)
-val secondaryContainerLight = Color(0xFFDAE2F9)
-val onSecondaryContainerLight = Color(0xFF131C2B)
-val tertiaryLight = Color(0xFF705575)
-val onTertiaryLight = Color(0xFFFFFFFF)
-val tertiaryContainerLight = Color(0xFFFAD8FD)
-val onTertiaryContainerLight = Color(0xFF28132E)
-val errorLight = Color(0xFFBA1A1A)
-val onErrorLight = Color(0xFFFFFFFF)
-val errorContainerLight = Color(0xFFFFDAD6)
-val onErrorContainerLight = Color(0xFF410002)
-val backgroundLight = Color(0xFFF9F9FF)
-val onBackgroundLight = Color(0xFF191C20)
-val surfaceLight = Color(0xFFF9F9FF)
-val onSurfaceLight = Color(0xFF191C20)
-val surfaceVariantLight = Color(0xFFE0E2EC)
-val onSurfaceVariantLight = Color(0xFF44474E)
-val outlineLight = Color(0xFF74777F)
-val outlineVariantLight = Color(0xFFC4C6D0)
-val scrimLight = Color(0xFF000000)
-val inverseSurfaceLight = Color(0xFF2E3036)
-val inverseOnSurfaceLight = Color(0xFFF0F0F7)
-val inversePrimaryLight = Color(0xFFAAC7FF)
-
-val primaryDark = Color(0xFFAAC7FF)
-val onPrimaryDark = Color(0xFF0A305F)
-val primaryContainerDark = Color(0xFF284777)
-val onPrimaryContainerDark = Color(0xFFD6E3FF)
-val secondaryDark = Color(0xFFBEC6DC)
-val onSecondaryDark = Color(0xFF283141)
-val secondaryContainerDark = Color(0xFF3E4759)
-val onSecondaryContainerDark = Color(0xFFDAE2F9)
-val tertiaryDark = Color(0xFFDDBCE0)
-val onTertiaryDark = Color(0xFF3F2844)
-val tertiaryContainerDark = Color(0xFF573E5C)
-val onTertiaryContainerDark = Color(0xFFFAD8FD)
-val errorDark = Color(0xFFFFB4AB)
-val onErrorDark = Color(0xFF690005)
-val errorContainerDark = Color(0xFF93000A)
-val onErrorContainerDark = Color(0xFFFFDAD6)
-val backgroundDark = Color(0xFF111318)
-val onBackgroundDark = Color(0xFFE2E2E9)
-val surfaceDark = Color(0xFF111318)
-val onSurfaceDark = Color(0xFFE2E2E9)
-val surfaceVariantDark = Color(0xFF44474E)
-val onSurfaceVariantDark = Color(0xFFC4C6D0)
-val outlineDark = Color(0xFF8E9099)
-val outlineVariantDark = Color(0xFF44474E)
-val scrimDark = Color(0xFF000000)
-val inverseSurfaceDark = Color(0xFFE2E2E9)
-val inverseOnSurfaceDark = Color(0xFF2E3036)
-val inversePrimaryDark = Color(0xFF415F91)
