--- app/src/main/java/com/paixao/dev/mobile_design_system/ui/theme/Shapes.kt
@@ -1,13 +1,0 @@
-package com.paixao.dev.mobile_design_system.ui.theme
-
-import androidx.compose.material3.ShapeDefaults
-import androidx.compose.material3.Shapes
-
-
-val designSystemShapes = Shapes(
-    extraSmall = ShapeDefaults.ExtraSmall,
-    small = ShapeDefaults.Small,
-    medium = ShapeDefaults.Medium,
-    large = ShapeDefaults.Large,
-    extraLarge = ShapeDefaults.ExtraLarge
-)
