+++ gradle/libs.versions.toml
@@ -38,4 +38,5 @@ androidx-material3-android = { group = "androidx.compose.material3", name = "mat
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
 jetbrainsKotlinAndroid = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+androidLibrary = { id = "com.android.library", version.ref = "agp" }
