+++ mdsfoundation/proguard-rules.pro
