+++ mdsfoundation/src/main/java/com/paixao/dev/mdsfoundation/theme/Theme.kt
@@ -1,0 +1,57 @@
+package com.paixao.dev.mdsfoundation.theme
+
+import android.app.Activity
+import android.os.Build
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Shapes
+import androidx.compose.material3.Typography
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
+import com.paixao.dev.mdsfoundation.tokens.designSystemDarkScheme
+import com.paixao.dev.mdsfoundation.tokens.designSystemLightScheme
+import com.paixao.dev.mdsfoundation.tokens.designSystemShapes
+import com.paixao.dev.mdsfoundation.tokens.designSystemTypography
+
+@Composable
+fun MobileDesignSystemTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    dynamicColor: Boolean = true,
+    darkColorScheme: ColorScheme = designSystemDarkScheme,
+    lightColorScheme:ColorScheme = designSystemLightScheme,
+    shapes: Shapes = designSystemShapes,
+    typography: Typography = designSystemTypography,
+    content: @Composable() () -> Unit
+) {
+    val colorScheme = when {
+        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+            val context = LocalContext.current
+            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+        }
+
+        darkTheme -> darkColorScheme
+        else -> lightColorScheme
+    }
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = colorScheme.primary.toArgb()
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
+        }
+    }
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        shapes = shapes,
+        typography = typography,
+        content = content
+    )
+}
