+++ mdsfoundation/src/main/java/com/paixao/dev/mdsfoundation/tokens/Shapes.kt
@@ -1,0 +1,12 @@
+package com.paixao.dev.mdsfoundation.tokens
+
+import androidx.compose.material3.ShapeDefaults
+import androidx.compose.material3.Shapes
+
+val designSystemShapes = Shapes(
+    extraSmall = ShapeDefaults.ExtraSmall,
+    small = ShapeDefaults.Small,
+    medium = ShapeDefaults.Medium,
+    large = ShapeDefaults.Large,
+    extraLarge = ShapeDefaults.ExtraLarge
+)
