--- mdstokens/src/main/java/com/paixao/dev/mds/tokens/tokens/Shapes.kt
@@ -1,12 +1,0 @@
-package com.paixao.dev.mds.tokens.tokens
-
-import androidx.compose.material3.ShapeDefaults
-import androidx.compose.material3.Shapes
-
-val designSystemShapes = Shapes(
-    extraSmall = ShapeDefaults.ExtraSmall,
-    small = ShapeDefaults.Small,
-    medium = ShapeDefaults.Medium,
-    large = ShapeDefaults.Large,
-    extraLarge = ShapeDefaults.ExtraLarge
-)
